Refactor: extract the function getTopUser on _helpers
