+++ _helpers.js
@@ -1,5 +1,7 @@
 const imgur = require('imgur-node-api')
-
+const db = require('./models')
+const User = db.User
+const sequelize = require('sequelize')
 function ensureAuthenticated(req) {
   return req.isAuthenticated();
 }
@@ -19,8 +21,42 @@ const imgPromise = (file) => {
   })
 }
 
+const getTopUser = async (req) => {
+  let users = await User.findAll({
+    limit: 10,
+    attributes: {
+      include: [
+        [
+          sequelize.literal(`(
+              SELECT COUNT(*)
+              FROM Followships AS Followship
+              WHERE Followship.followingId = User.id
+            )`),
+          'FollowerCount'
+        ]
+      ]
+    },
+    order: [
+      [sequelize.literal('FollowerCount'), 'DESC']
+    ],
+    where: {
+      role: 'user'
+    },
+    include: [
+      { model: User, as: 'Followers' } //找出每個User被追蹤的名單(user.Followers)
+    ]
+  })
+
+  users = users.map(user => ({
+    ...user.dataValues,
+    isFollowed: getUser(req).Followings.some(d => d.id === user.id)
+  }))
+  return users
+}
+
 module.exports = {
   ensureAuthenticated,
   getUser,
-  imgPromise
+  imgPromise,
+  getTopUser
 };
