+++ controllers/userController.js
@@ -8,6 +8,7 @@ const Reply = db.Reply
 const Like = db.Like
 const sequelize = require('sequelize')
 const imgPromise = require('../_helpers').imgPromise
+const getTopUser = require('../_helpers').getTopUser
 if (process.env.NODE_ENV !== 'production') {
   require('dotenv').config()
 }
@@ -82,36 +83,8 @@ const userController = {
       ReplyCount: t.Replies.length,
       isLiked: helpers.getUser(req).Likes ? helpers.getUser(req).Likes.map(d => d.TweetId).includes(t.id) : false
     }))
-    //user data
-    let users = await User.findAll({
-      limit: 10,
-      attributes: {
-        include: [
-          [
-            sequelize.literal(`(
-              SELECT COUNT(*)
-              FROM Followships AS Followship
-              WHERE Followship.followingId = User.id
-            )`),
-            'FollowerCount'
-          ]
-        ]
-      },
-      order: [
-        [sequelize.literal('FollowerCount'), 'DESC']
-      ],
-      where: {
-        role: 'user'
-      },
-      include: [
-        { model: User, as: 'Followers' } //找出每個User被追蹤的名單(user.Followers)
-      ]
-    })
-
-    users = users.map(user => ({
-      ...user.dataValues,
-      isFollowed: helpers.getUser(req).Followings.some(d => d.id === user.id)
-    }))
+    //getTopUser
+    let users = await getTopUser(req)
 
     if (Number(helpers.getUser(req).id) !== Number(req.body.id)) {
       await Followship.create({
@@ -133,6 +106,7 @@ const userController = {
   },
 
   getUserPage: async (req, res) => {
+    //取userView是為了後續跟當前登入者user的id比對去決定是否顯示編輯個人資料
     let userView = await User.findByPk(req.params.id, {
       include: [
         { model: Like },
@@ -147,8 +121,10 @@ const userController = {
     const totalLikes = userView.Likes.length
     const totalFollowers = userView.Followers.length
     const totalFollowings = userView.Followings.length
+    //////////////////////////////user data to show top 10 user
+    let users = await getTopUser(req)
 
-    return res.render('user', { userView, totalReplies, totalLikes, totalFollowers, totalFollowings })
+    return res.render('user', { userView, totalReplies, totalLikes, totalFollowers, totalFollowings, users })
   },
 
   getUsers: async (req, res) => {
@@ -161,7 +137,8 @@ const userController = {
   },
 
   getUserTweetsAndRepliesPage: async (req, res) => {
-    //todo
+    //user data to show top 10 user
+    let users = await getTopUser(req)
   },
   getUserLikesPage: async (req, res) => {
     let userView = await User.findByPk(req.params.id, {
@@ -180,7 +157,10 @@ const userController = {
     })
     const totalFollowers = userView.Followers.length
     const totalFollowings = userView.Followings.length
-    return res.render('likes', { userView, totalFollowers, totalFollowings })
+
+    //user data to show top 10 user
+    let users = await getTopUser(req)
+    return res.render('likes', { userView, totalFollowers, totalFollowings, users })
   },
 
   editUserFromEditPage: async (req, res) => {
@@ -220,6 +200,7 @@ const userController = {
   },
 
   getUserFollowingPage: async (req, res) => {
+    //user following data
     let user = await User.findByPk(req.params.id, {
       attributes: {
         include: [
@@ -238,9 +219,12 @@ const userController = {
       ]
     })
     user = user.toJSON()
-    return res.render('userFollowing', { user })
+    //users data to show top user on right side
+    let users = await getTopUser(req)
+    return res.render('userFollowing', { user, users })
   },
   getUserFollowerPage: async (req, res) => {
+    //User follower Data
     let user = await User.findByPk(req.params.id, {
       attributes: {
         include: [
@@ -262,7 +246,9 @@ const userController = {
     user.Followers.map(user => {
       user.isFollowed = helpers.getUser(req).Followings.map(d => d.id).includes(user.id)
     })
-    return res.render('userFollower', { user })
+    //users data to show top user on right side
+    let users = await getTopUser(req)
+    return res.render('userFollower', { user, users })
   }
 }
